code restructure
Move the file specific codes to file.go, move some generic logic to utils.go
